requirements.txt
watchdog>=3.0.0
anthropic>=0.18.0
chromadb>=0.4.22
pyyaml>=6.0
ruamel.yaml>=0.18.0
python-dotenv>=1.0.0
